refactor: 영속성 컨텍스트 특징 유의하여 서비스 레이어 리팩토링

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -11,6 +11,7 @@ import com.sprint.mission.discodeit.entity.ReadStatus;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -20,14 +21,17 @@ import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class BasicChannelService implements ChannelService {
 
   private final ChannelRepository channelRepository;
   private final ReadStatusRepository readStatusRepository;
   private final MessageRepository messageRepository;
+  private final UserRepository userRepository;
 
   @Override
   public Channel create(PublicChannelCreateRequest channelCreateRequest) {
@@ -43,18 +47,21 @@ public class BasicChannelService implements ChannelService {
     Channel channel = new Channel(ChannelType.PRIVATE, null, null);
     Channel createdChannel = channelRepository.save(channel);
 
-    channelCreateRequest.participantIds().stream() // 채널의 참여자들을 스트림으로 변환
-        .map(userId -> new ReadStatus(userId, createdChannel.getId(),
-            channel.getCreatedAt())) // 각 참여자마다 readStatus 객체 생성
-        .forEach(readStatusRepository::save);
+    List<ReadStatus> readStatusList = userRepository.findAllById(
+            channelCreateRequest.participantIds())
+        .stream()
+        .map(user -> new ReadStatus(user, createdChannel, createdChannel.getCreatedAt()))
+        .toList();
 
+    readStatusRepository.saveAll(readStatusList);
     return createdChannel;
   }
 
+  @Transactional(readOnly = true)
   @Override
   public ChannelDTO find(UUID channelId) {
     return channelRepository.findById(channelId)
-        .map(channel -> toChannelDTO(channel))
+        .map(this::toChannelDTO)
         .orElseThrow(() -> new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다."));
   }
 
@@ -71,24 +78,25 @@ public class BasicChannelService implements ChannelService {
     }
 
     channel.update(newName, newDescription);
-    return channelRepository.save(channel);
+    return channel;
   }
 
   @Override
   public void delete(UUID channelId) {
-    Channel channel = channelRepository.findById(channelId)
-        .orElseThrow(() -> new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다."));
+    if (!channelRepository.existsById(channelId)) {
+      throw new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다.");
+    }
 
-    messageRepository.deleteAllByChannelId(channel.getId());
-    readStatusRepository.deleteAllByChannelId(channel.getId());
-
+    messageRepository.deleteAllByChannelId(channelId);
+    readStatusRepository.deleteAllByChannelId(channelId);
     channelRepository.deleteById(channelId);
   }
 
+  @Transactional(readOnly = true)
   @Override
   public List<ChannelDTO> findAllByUserId(UUID userId) {
     List<UUID> mySubscribedChannelIds = readStatusRepository.findAllByUserId(userId).stream()
-        .map(ReadStatus::getChannelId)
+        .map(readStatus -> readStatus.getChannel().getId())
         .toList();
 
     return channelRepository.findAll().stream()
@@ -115,10 +123,10 @@ public class BasicChannelService implements ChannelService {
 
     List<UUID> participantIds = new ArrayList<>();
     if (channel.getType().equals(ChannelType.PRIVATE)) {
-      readStatusRepository.findAllByChannelId(channel.getId())
+      participantIds = readStatusRepository.findAllByChannelId(channel.getId())
           .stream()
-          .map(ReadStatus::getUserId)
-          .forEach(participantIds::add);
+          .map(readStatus -> readStatus.getUser().getId())
+          .toList();
     }
 
     return ChannelDTO.fromEntity(channel, lastMessageAt, participantIds);
